LayoutTests/fast/css/getComputedStyle/getComputedStyle-height.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<style type="text/css">
    table td {
        padding: 20px;
        border: 10px solid blue;
    }

    #cell1 {
        height: 200px;
    }
</style>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<table border="0" cellspacing="0" cellpadding="0" id="table">
    <tr id="row1">
        <td id="cell1">height</td>
    </tr>
</table>
<script>

description("Test the computed height of a cell : https://bugs.webkit.org/show_bug.cgi?id=33593.")

e = document.getElementById('cell1');
table = document.getElementById('table');
computedStyle = window.getComputedStyle(e, null);

shouldBe("computedStyle.getPropertyValue('height')", "'200px'");

document.body.removeChild(table);

</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
