third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-height-expected.txt

Test the computed height of a cell : https://bugs.webkit.org/show_bug.cgi?id=33593.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('height') is '200px'
PASS successfullyParsed is true

TEST COMPLETE
